pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "lastpass-py"
dynamic = ["version"]
description = "Complete Python implementation of LastPass CLI with API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "GPL-2.0-or-later"}
authors = [
    {name = "LastPass Python Contributors"}
]
keywords = ["lastpass", "password", "manager", "cli", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.28.0",
    "pycryptodome>=3.15.0",
]

[project.optional-dependencies]
clipboard = ["pyperclip>=1.8.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "responses>=0.22.0",
    "coverage>=7.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=0.950",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "responses>=0.22.0",
    "coverage>=7.0.0",
]

[project.scripts]
lpass = "lastpass.cli:main"

[project.urls]
Homepage = "https://github.com/dynacylabs/lastpass-py"
Documentation = "https://github.com/dynacylabs/lastpass-py/blob/main/README.md"
Repository = "https://github.com/dynacylabs/lastpass-py"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=lastpass --cov-report=term-missing --strict-markers --tb=short"
markers = [
    "unit: Unit tests with mocked dependencies",
    "integration: Integration tests that hit real external services",
    "slow: Tests that are slow to run",
    "live: Tests requiring live API connection (deselect with '-m \"not live\"')",
]

[tool.coverage.report]
fail_under = 95
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.setuptools_scm]
write_to = "lastpass/_version.py"
